+++ source/BilliardsSDL/Game/Scenes/Scene.h
@@ -25,8 +25,8 @@ public:
 
 protected:
 	virtual void CreateGameObjects() = 0;
-	GameObject* CreateGameObject(const Vector2<int>& position);
-	std::shared_ptr<Image> CreateImageComponent(GameObject* owner, const ImageResourceData& resourceData, const Vector2<int>& size);
+	GameObject* CreateGameObject(const Vector2<float>& position);
+	std::shared_ptr<Image> CreateImageComponent(GameObject* owner, const ImageResourceData& resourceData, const Vector2<float>& size);
 	std::shared_ptr<Text> CreateTextComponent(GameObject* owner, const TextResourceData& resourceData,
 											  const std::string& textString, const int pointSize);
